Move Library/Manager to new namespace.
