+++ src/main/java/net/glowstone/GlowServer.java
@@ -122,8 +122,8 @@ import net.glowstone.scoreboard.GlowScoreboardManager;
 import net.glowstone.util.GlowHelpMap;
 import net.glowstone.util.GlowServerIcon;
 import net.glowstone.util.GlowUnsafeValues;
-import net.glowstone.util.Library;
-import net.glowstone.util.LibraryManager;
+import net.glowstone.library.Library;
+import net.glowstone.library.LibraryManager;
 import net.glowstone.util.OpenCompute;
 import net.glowstone.util.SecurityUtils;
 import net.glowstone.util.ShutdownMonitorThread;
