+++ src/main/java/net/glowstone/library/Library.java
@@ -1,9 +1,9 @@
-package net.glowstone.util;
+package net.glowstone.library;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Objects;
-import net.glowstone.util.LibraryManager.HashAlgorithm;
+import net.glowstone.library.LibraryManager.HashAlgorithm;
 
 public class Library {
     private static final String ARTIFACT_ID_KEY = "artifact-id";
