+++ src/main/java/net/glowstone/util/config/ServerConfig.java
@@ -18,8 +18,8 @@ import java.util.logging.Level;
 import lombok.Getter;
 import net.glowstone.GlowServer;
 import net.glowstone.util.DynamicallyTypedMap;
-import net.glowstone.util.Library;
-import net.glowstone.util.LibraryManager.HashAlgorithm;
+import net.glowstone.library.Library;
+import net.glowstone.library.LibraryManager.HashAlgorithm;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.YamlConfiguration;
